requirements.txt
fastapi
uvicorn[standard]
requests
scikit-learn
pandas
python-dotenv
httpx
stripe
